server/requirements.txt
flask
flask-sock
google-genai
numpy
librosa
soundfile
